+++ src/collective/nitf/config.py
@@ -191,5 +191,6 @@ URGENCIES = SimpleVocabulary([
 
 # defaults
 DEFAULT_GENRE = u'Current'
+DEFAULT_SECTION = _(u'General')
 DEFAULT_URGENCY = NORMAL
 DEFAULT_RELATABLE_CONTENT_TYPES = [u'collective.nitf.content']